# [Backport 6.0] replica: remove rwlock for protecting iteration over storage group map