--- a/test/perf/perf_fast_forward.cc
+++ b/test/perf/perf_fast_forward.cc
@@ -1744,6 +1744,14 @@ replica::table& find_table(replica::database& db, dataset& ds) {
     return db.find_column_family("ks", ds.table_name());
 }
 
+static std::vector<replica::memtable*> active_memtables(replica::table& t) {
+    std::vector<replica::memtable*> active_memtables;
+    t.for_each_active_memtable([&] (replica::memtable& mt) {
+        active_memtables.push_back(&mt);
+    });
+    return active_memtables;
+}
+
 static
 void populate(const std::vector<dataset*>& datasets, cql_test_env& env, const table_config& cfg, size_t flush_threshold) {
     drop_keyspace_if_exists(env, "ks");
@@ -1775,7 +1783,7 @@ void populate(const std::vector<dataset*>& datasets, cql_test_env& env, const ta
                 auto gen = ds.make_generator(s, cfg);
                 while (auto mopt = gen()) {
                     ++fragments;
-                    replica::memtable& active_memtable = *cf.active_memtables().front();
+                    replica::memtable& active_memtable = *active_memtables(cf).front();
                     active_memtable.apply(*mopt);
                     if (active_memtable.region().occupancy().used_space() > flush_threshold) {
                         metrics_snapshot before;