@scylladb-promoter 2024-07-20
#### :red_circle: CI State: FAILURE

:white_check_mark: - [Build](https://jenkins.scylladb.com/job/scylla-6.0/job/scylla-ci/143/)
:x: - [Unit Tests Custom](https://jenkins.scylladb.com/job/scylla-6.0/job/scylla-ci/143/)
**The following new/updated tests ran 100 times for each mode:**
:small_blue_diamond: boost/commitlog_test
:small_blue_diamond: boost/memtable_test
:small_blue_diamond: boost/mutation_test
:small_blue_diamond: topology_experimental_raft/test_tablets
#### Failed Tests (1/6733):
- [test_topology_changes](https://jenkins.scylladb.com//job/scylla-6.0/job/scylla-ci/143/testReport/junit/%28root%29/test_tablets/test_topology_changes) [:mag:](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_topology_changes)
#### Build Details:
- Duration: 5 hr 47 min
- Builder: spider4.cloudius-systems.com