@bhalevy 2024-07-21
@raphaelsc please investigate unit test failure