@raphaelsc 2024-08-01
this series must include https://github.com/scylladb/scylladb/pull/19987