@raphaelsc 2024-08-06
> Guys, this PR is critical for 2024.2 cc @mykaul

we're waiting on a complementary fix to land on master: https://github.com/scylladb/scylladb/pull/19987