@raphaelsc 2024-08-13
v2: included fix for race between split compaction and migration (https://github.com/scylladb/scylladb/pull/19987)