@scylladb-promoter 2024-08-14
#### :green_circle: CI State: SUCCESS

:white_check_mark: - [Build](https://jenkins.scylladb.com/job/scylla-6.0/job/scylla-ci/176/)
:white_check_mark: - [Unit Tests Custom](https://jenkins.scylladb.com/job/scylla-6.0/job/scylla-ci/176/)
**The following new/updated tests ran 100 times for each mode:**
:small_blue_diamond: boost/commitlog_test
:small_blue_diamond: boost/memtable_test
:small_blue_diamond: boost/mutation_test
:small_blue_diamond: topology_experimental_raft/test_tablets
:white_check_mark: - [Container Test](https://jenkins.scylladb.com/job/scylla-6.0/job/scylla-ci/176/)
:white_check_mark: - [dtest with topology changes](https://jenkins.scylladb.com/job/scylla-6.0/job/gating-dtest-release-with-consistent-topology-changes/293/)
:white_check_mark: - [dtest](https://jenkins.scylladb.com/job/scylla-6.0/job/gating-dtest-release/295/)
:white_check_mark: - [Unit Tests](https://jenkins.scylladb.com/job/scylla-6.0/job/scylla-ci/176/)
#### Build Details:
- Duration: 7 hr 34 min
- Builder: spider1.cloudius-systems.com